backend/requirements.txt
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
gunicorn==22.0.0
psycopg[binary]==3.2.13
python-dotenv==1.0.1
pydantic==2.8.2
PyPDF==4.3.1
python-docx==1.1.2
requests==2.32.3
